Refactor: Generalize buffer allocator for pooled upload buffers
This patch refactors the `BucketizedBufferAllocator` to make it more
generic and uses it to add a dedicated allocator for upload buffers.

The `BucketizedBufferAllocator` is modified to accept a `std::function`
for resource creation instead of a `Device*`. This decouples the
  allocator from the `Device` class and allows it to be reused for
  different heap types.

A new `m_uploadAllocator` is added to the `Device` class, configured to
create pooled buffers in the UPLOAD heap. The `Device::Upload` method is
updated to use this new allocator, reducing the overhead of creating and
destroying temporary upload buffers.

Additionally, explicit `ExecuteCommandList()` calls are removed from
`Device::Upload` and `primitives::copy` to allow for better command
batching, improving GPU submission efficiency.