perf(dml): Restore fast copy implementation using CopyBufferRegion
The use of `DML_ELEMENT_WISE_IDENTITY` to make the copy primitive
graph-recordable resulted in a severe performance drawback.

This commit reverts the copy implementation back to using the native
D3D12 `CopyBufferRegion` command, as it is substantially faster for this
task.

This fixes the performance regression at the cost of breaking graph
recording for operations that include a copy. This trade-off is
acceptable for now to restore baseline performance.